Specify options for a musicassessr test.

Specify options for a musicassessr test.

Usage

musicassessr_opt(
  setup_pages = TRUE,
  setup_options = setup_pages_options(),
  use_musicassessr_db = FALSE,
  app_name = "",
  midi_file_playback = FALSE,
  visual_notation = FALSE,
  record_audio = TRUE,
  midi_input = FALSE,
  fake_range = TRUE,
  musicassessr_aws = FALSE,
  experiment_id = NULL,
  experiment_condition_id = NULL,
  user_id = NULL,
  instrument_id = NULL,
  get_p_id = FALSE
)

Arguments

setup_pages

Should there be setup pages?

setup_options

Setup page options.

use_musicassessr_db

Is a SQL backend being used?

app_name

App name for audio recording apps.

midi_file_playback

Will MIDI files need to be played back?

visual_notation

Will there be visual notation?

record_audio

Is audio recording required?

midi_input

Is MIDI input required?

fake_range

Should the vocal/instrument range be faked?

musicassessr_aws

Is this being run on Amazon Web Services (AWS)?

experiment_id

An experiment ID.

experiment_condition_id

An experiment condition ID.

user_id

A user ID.

instrument_id

An instrument ID.

get_p_id

Should a participant ID get collected at the beginning of the test?