Convert a melody from relative to absolute form by centering the mean of the stimuli on mean of the user's range.

Convert a melody from relative to absolute form by centering the mean of the stimuli on mean of the user's range.

Usage

rel_to_abs_mel_mean_centred(
  rel_melody,
  bottom_range,
  top_range,
  plot = FALSE,
  range = NULL,
  transpose = NULL
)

Arguments

transpose