Setup page options for a musicassessr test

Setup page options for a musicassessr test

Usage

setup_pages_options(
  input_type = c("microphone", "midi_keyboard", "midi_keyboard_and_microphone",
    "midi_keyboard_or_microphone"),
  headphones = TRUE,
  SNR_test = TRUE,
  min_SNR = 14,
  get_user_info = TRUE,
  demo = FALSE,
  get_instrument_range = FALSE,
  absolute_url = "https://www.adaptiveeartraining.com",
  select_instrument = FALSE,
  get_instrument_range_musical_notation = FALSE,
  adjust_range = FALSE,
  test_type = c("voice", "instrument"),
  microphone_test = TRUE,
  allow_repeat_SNR_tests = TRUE,
  report_SNR = FALSE,
  concise_wording = TRUE,
  skip_setup = FALSE,
  get_self_chosen_anonymous_id = FALSE,
  musical_instrument = FALSE,
  fake_range = FALSE
)

Arguments

input_type

The form of musicassessr test.

headphones

Deploy a headphone setup page.

SNR_test

Deploy an SNR test.

min_SNR

The minimum accepted SNR value.

get_user_info

Get user info from their browser.

demo

Deploy demo test.

get_instrument_range

Get the user's pitch/instrument range at test time.

absolute_url

If in production, what is the absolute URL.

select_instrument

Deploy instrument selection page.

get_instrument_range_musical_notation

When displaying instrument range, provide visual music notation.

adjust_range

Adjust range.

test_type

Voice vs. instrument test.

microphone_test

Deploy microphone test page.

allow_repeat_SNR_tests

Allow repeated SNR tests, if FALSE, then participant only gets one go and the test will fail if the SNR test fails.

report_SNR

Should the SNR be reported to the user.

concise_wording

Should the wording used by the concise version for not.

skip_setup

Whether to skip setup. Can be TRUE (skip whole setup), FALSE or "except_microphone" (only setup the microphone but no other steps).

get_self_chosen_anonymous_id

Whether to ask participant to provide an anonymous ID.

musical_instrument

Whether the participant is required to have a musical instrument.

fake_range

Should a range be faked?